wait after boot lmk

+++ zip/system/etc/parrotmod/postboot.sh
@@ -3,10 +3,21 @@ bb=/system/etc/parrotmod/busybox
 
 while [ "$(getprop sys.boot_completed)" = "" ]; do sleep 1; done
 
-settings put global sys_storage_full_threshold_bytes 8388608
-settings put global sys_storage_threshold_percentage 2
-settings put global sys_storage_threshold_max_bytes 104857600
+if [ "$(cat /data/lastpmver_univ.txt)" != "1" ]; then
 
-echo 1 > /data/lastpmver_univ.txt
+  settings put global sys_storage_full_threshold_bytes 8388608
+  settings put global sys_storage_threshold_percentage 2
+  settings put global sys_storage_threshold_max_bytes 104857600
+  
+  echo 1 > /data/lastpmver_univ.txt
+  
+  am start -a android.intent.action.REBOOT # cleaner reboot
+  
+fi
 
-am start -a android.intent.action.REBOOT # cleaner reboot
+# http://review.cyanogenmod.org/#/c/101476/ instead ???
+$bb chmod -R 0775 /sys/module/lowmemorykiller/parameters
+echo "0,1,2,5,7,16" > /sys/module/lowmemorykiller/parameters/adj 
+echo "9933,10728,14950,17510,20019,31385" > /sys/module/lowmemorykiller/parameters/minfree 
+echo "24" > /sys/module/lowmemorykiller/parameters/cost # default 32
+$bb chmod -R 0555 /sys/module/lowmemorykiller/parameters # so android can't edit it
